% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alokasi_dt}
\alias{alokasi_dt}
\title{Example of Allocation Data}
\format{
\subsection{\code{alokasi_dt}}{

A data frame with 34 rows and 3 columns:
\describe{
\item{kdprov}{province code}
\item{jml_kabkota}{Population or number of regencies/cities}
\item{n_primary}{Sample Allocation}
...
}
}
}
\usage{
alokasi_dt
}
\description{
Example of Allocation Data for Sampling Purposes
}
\keyword{datasets}
