% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingin.R
\name{get_allocation}
\alias{get_allocation}
\title{Get allocation from population tabulation given total allocation}
\usage{
get_allocation(data, alokasi, group, pop_var = "jml", secondary = 0)
}
\arguments{
\item{data}{population tabulation dataframe}

\item{alokasi}{total allocation dataframe}

\item{group}{group of allocation level to be obtained}

\item{pop_var}{population variable in data}

\item{secondary}{how many times the secondary sample compares to primary sample}
}
\value{
allocation at more detailed level
}
\description{
Get allocation from population tabulation given total allocation
}
\examples{
\dontrun{
alokasi_dt = get_allocation(
   data = contoh_alokasi \%>\% mutate(nasional = 1)
   , alokasi = 100
   , group = c("nasional")
   , pop_var = "jml_kabkota"
) \%>\%
select(-nasional)
}
}
