% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_dt}
\alias{pop_dt}
\title{Indonesian Population (SP2020)}
\format{
\subsection{\code{pop_dt}}{

A data frame with 514 rows and 8 columns:
\describe{
\item{idkab}{region id}
\item{kdprov}{province code}
\item{kdkab}{regency/city code}
\item{nmprov}{province name}
\item{nmkab}{regency/city name}
\item{Laki-laki}{Male Population}
\item{Perempuan}{Female Population}
\item{Total}{Total Population}
...
}
}
}
\source{
\url{https://sensus.bps.go.id/main/index/sp2020}
}
\usage{
pop_dt
}
\description{
Tabulation of Indonesia's population based on the results of the 2020 population census by regency/city and gender
}
\keyword{datasets}
