% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingin.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{round_preserve_sum}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{a number}

\item{digits}{0 (default)}
}
\description{
round_preserve_sum
}
