% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingin.R
\name{doSampling}
\alias{doSampling}
\title{Select samples given its parameters}
\usage{
doSampling(
  pop,
  alloc,
  nsampel,
  type,
  strata = NULL,
  ident = c("kdprov", "kdkab"),
  method = "systematic",
  auxVar = NA,
  seed = 1
)
}
\arguments{
\item{pop}{pop dataframe}

\item{alloc}{allocation dataframe}

\item{nsampel}{variable on alloc df as allocation sample}

\item{type}{type value for sample classification ('U' = Utama, 'P' = pengganti)}

\item{strata}{strata variable, must available on both pop and alloc dataframe}

\item{ident}{group by on allocation dataframe}

\item{method}{method of sampling : \code{"systematic"} (the default) or \code{"pps"}}

\item{auxVar}{auxiliary variable for pps sampling (\code{method = "pps"})}

\item{seed}{seed}
}
\value{
list of pop (\code{"pop"}), samples data (\code{"dsampel"}), and details (\code{"rincian"})
}
\description{
Select samples given its parameters
}
\examples{
\dontrun{
library(samplingin)

dtSampling_pps = doSampling(
   pop = pop_dt
   , alloc = alokasi_dt
   , nsampel = "n_primary"
   , type = "U"
   , ident = c("kdprov")
   , method = "pps"
   , auxVar = "Total"
   , seed = 1234
)

pop_dt = dtSampling_pps$pop
dsampel = dtSampling_pps$dsampel
rincian = dtSampling_pps$rincian


dtSampling_sys = doSampling(
   pop = pop_dt
   , alloc = alokasi_dt
   , nsampel = "n_primary"
   , type = "U"
   , ident = c("kdprov")
   , method = "systematic"
   , seed = 4321
)

pop_dt = dtSampling_sys$pop
dsampel = dtSampling_sys$dsampel
rincian = dtSampling_sys$rincian
}
}
